191113 수업내용(Hashtag, Social Login)

+++ django_form/articles/views.py
@@ -1,16 +1,21 @@
-from django.shortcuts import render, redirect, get_object_or_404,HttpResponse
-from .models import Article, Comment
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
+from .models import Article, Comment, Hashtag
 from .forms import ArticleForm, CommentForm
 from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
-# from IPython import embed
+from IPython import embed
+import hashlib
+from django.contrib.auth import get_user_model
 
 # Create your views here.
 def index(request):
-    # embed()
     articles = Article.objects.all()
-    context = {'articles': articles}
+    context = {
+        'articles': articles,
+    }
     return render(request, 'articles/index.html', context)
+
 
 # 이런 식으로도 사용 가능하다! @login_required(login_url="/accounts/test/")
 @login_required
@@ -20,23 +25,32 @@ def create(request):
         form = ArticleForm(request.POST)
         # 폼이 유효한지 체크한다.
         if form.is_valid():
-            article = form.save()
-            # title = form.cleaned_data.get("title")
-            # content = form.cleaned_data.get("content")
-            # article = Article(title=title, content=content)
-            # article.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
+            #Hashtag
+            for word in article.content.split():
+                if word.startswith("#"):
+                    hashtag, created= Hashtag.objects.get_or_create(content=word)
+                    article.hashtags.add(hashtag)
             return redirect("articles:detail", article.pk)
     else:
         form = ArticleForm()
     context = {'form': form}
     return render(request, 'articles/create.html', context)
 
+
 def detail(request, article_pk):
     # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
     comments = Comment.objects.filter(article=article)
+    person= get_object_or_404(get_user_model(), pk=article.user_id)
     comment_form = CommentForm()
-    context = {'article': article, 'comment_form': comment_form, 'comments': comments}
+    context = {'article': article,
+               'comment_form': comment_form, 
+               'comments': comments,
+               'person':person,
+               }
     return render(request, 'articles/detail.html', context)
 
 
@@ -44,27 +58,32 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if request.user == article.user:
+            article.delete()
         return redirect("articles:index")
     return HttpResponse('You are Unauthorized', status=401)
+
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == "POST":
-        form = ArticleForm(request.POST)
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            # article.title = form.cleaned_data.get("title")
-            # article.content = form.cleaned_data.get("content")
-            # article.save()
-            return redirect("articles:detail", article_pk)
+    if request.user == article.user:
+        if request.method == "POST":
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                article= form.save()
+                article.hashtags.clear()
+                for word in article.content.split():
+                    if word.startswith("#"):
+                        hashtag, created = Hashtag.objects.get_or_create(contente=word)
+                        article.hashtags.add(hashtag)
+                return redirect("articles:detail", article_pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect("articles:index")
     context = {'form': form}
     return render(request, 'articles/create.html', context)
-
 
 @require_POST
 def comments_create(request, article_pk):
@@ -73,16 +92,52 @@ def comments_create(request, article_pk):
         comment_form = CommentForm(request.POST)
         if comment_form.is_valid():
             comment = comment_form.save(commit=False)
+            comment.user = request.user # 댓글 작성한 사용자 저장
             comment.article = article
             comment.save()
         return redirect("articles:detail", article_pk)
     return HttpResponse('You are Unauthorized', status=401)
 
-
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
-        return redirect("articles:detail", article_pk)
-    return HttpResponse('You are Unauthorized', status=401)
+        if request.user == comment.user:
+            comment.delete()
+    return redirect("articles:detail", article_pk)
+
+
+def like(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    user = request.user # 요청을 보낸 유저
+    
+    # 해당 게시글에 좋아요를 누른 사람들 중에
+    # user.pk를 가진 유저가 존재하면,
+    if request.user.is_authenticated:
+        if article.like_users.filter(pk=user.pk).exists():
+            # user를 삭제하고 (좋아요를 취소)
+            article.like_users.remove(user)
+        else:
+            article.like_users.add(user)
+        return redirect("articles:index")
+    return redirect("accounts:login")
+
+
+@login_required
+def follow(request, article_pk, user_pk):
+    #게시글 유저
+    you = get_object_or_404(get_user_model(), pk=user_pk)
+    #접속 유저
+    me= request.user
+    if you != me:
+        if you.followers.filter(pk=me.pk).exists():
+            you.followers.remove(me)
+        else:
+            you.followers.add(me)
+    return redirect('articles:detail', article_pk)
+
+def hashtag(request, hash_pk):
+    hashtag= get_object_or_404(Hashtag, pk=hash_pk)
+    articles= hashtag.article_set.order_by("-pk")
+    context={'hashtag':hashtag, 'articles':articles}
+    return render(request, 'articles/hashtag.html', context)
